Separate server.js in two files
